added "auto_change_to_camera", and spread across files
